Expose signing key id for service account creds
The id of the service account key used to sign JWTs to request IAP
access tokens is also encoded in the JWT headers. This is not the
prettiest way to standardize reading this value from the keyfile json,
but it's better than reading it outside of the ServiceCredentials class.
The whole structure of the key / signer handling in this gem probably
wants a reevaluation if the intent will be to continue using it for IAP
tokens. The google-auth-library-python package could serve as a good
guide for making the JWT handling a little cleaner. In particular, it'd
be great to get rid of the 4-element array response from the
read_json_io module method.